refactored actions folder
* removed constant file
* relocated action type enums to index folder
