--- src/actions/constants.ts
@@ -1,9 +1,0 @@
-enum SquareActionTypes {
-  GET_ALL_SQUARES = 'GET_ALL_SQUARES',
-  GET_SQUARE = 'GET_SQUARE',
-  CREATE_SQUARE_SUCCEEDED = 'CREATE_SQUARE_SUCCEEDED',
-  DELETE_SQUARE_SUCCEEDED = 'DELETE_SQUARE_SUCCEEDED',
-  UPDATE_SQUARE_SUCCEEDED = 'UPDATE_SQUARE_SUCCEEDED',
-}
-
-export default SquareActionTypes;
