+++ src/actions/index.ts
@@ -1,5 +1,14 @@
-import SquareActionTypes from './constants';
 import SquareInterface from '../components/SquareInterface';
+
+enum SquareActionTypes {
+  GET_ALL_SQUARES = 'GET_ALL_SQUARES',
+  GET_SQUARE = 'GET_SQUARE',
+  CREATE_SQUARE_SUCCEEDED = 'CREATE_SQUARE_SUCCEEDED',
+  DELETE_SQUARE_SUCCEEDED = 'DELETE_SQUARE_SUCCEEDED',
+  UPDATE_SQUARE_SUCCEEDED = 'UPDATE_SQUARE_SUCCEEDED',
+}
+
+export default SquareActionTypes;
 
 export interface IGetAllSquaresSucceeded {
   readonly type: SquareActionTypes.GET_ALL_SQUARES;
